Fix normalMapScale to UnpacknormalScale function

+++ com.microsoft.mrtk.graphicstools.unity/Runtime/Shaders/GraphicsToolsStandardInput.hlsl
@@ -198,7 +198,7 @@ CBUFFER_START(UnityPerMaterial)
     half _Metallic;
     half _Smoothness;
 
-    float _NormalMapScale;
+    half _NormalMapScale;
     // #if defined(_ALPHA_CLIP)
     half _Cutoff;
